Refactor code by exporting YouTube related functions to a new class
The YouTube related functionalities previously in 'app.py' have been encapsulated within a new YouTube class in the 'youtube.py' file. The aim of this refactoring is to separate concerns, leading to a cleaner and more manageable structure. The 'app.py' file's YouTube related functions were commented out rather than deleted.
